+++ sdk/servicebus/service-bus/samples/javascript/advanced/managementClient.js
@@ -6,12 +6,12 @@
   For samples using the current stable version of the package, please use the link below:
   https://github.com/Azure/azure-sdk-for-js/blob/master/sdk/servicebus/service-bus/samples-v1
   
-  This sample demonstrates how the ServiceBusManagementClient can be used to manage the resources of a service bus namespace.
+  This sample demonstrates how the ServiceBusAdministrationClient can be used to manage the resources of a service bus namespace.
 
   See https://docs.microsoft.com/rest/api/servicebus/resource-provider-apis to learn more.
 */
 
-const { ServiceBusManagementClient } = require("@azure/service-bus");
+const { ServiceBusAdministrationClient } = require("@azure/service-bus");
 
 // Load the .env file if it exists
 require("dotenv").config();
@@ -23,29 +23,29 @@ const queueName = process.env.QUEUE_NAME || "<queue name>";
 async function main() {
   // You can also use AAD credentials from `@azure/identity` along with the host url
   // instead of the connection string for authentication.
-  const serviceBusManagementClient = new ServiceBusManagementClient(connectionString);
+  const ServiceBusAdministrationClient = new ServiceBusAdministrationClient(connectionString);
 
   // Similarly, you can create topics and subscriptions as well.
-  const createQueueResponse = await serviceBusManagementClient.createQueue(queueName);
+  const createQueueResponse = await ServiceBusAdministrationClient.createQueue(queueName);
   console.log("Created queue with name - ", createQueueResponse.name);
 
-  const getQueueResponse = await serviceBusManagementClient.getQueue(queueName);
+  const getQueueResponse = await ServiceBusAdministrationClient.getQueue(queueName);
   console.log("(Current)max delivery count = ", getQueueResponse.maxDeliveryCount);
 
   getQueueResponse.maxDeliveryCount = 12;
-  const updateQueueResponse = await serviceBusManagementClient.updateQueue(getQueueResponse);
+  const updateQueueResponse = await ServiceBusAdministrationClient.updateQueue(getQueueResponse);
   console.log("(Updated)max delivery count = ", updateQueueResponse.maxDeliveryCount);
 
-  const queueRuntimeProperties = await serviceBusManagementClient.getQueueRuntimeProperties(
+  const queueRuntimeProperties = await ServiceBusAdministrationClient.getQueueRuntimeProperties(
     queueName
   );
   console.log("Number of messages in the queue = ", queueRuntimeProperties.totalMessageCount);
 
-  const namespaceInfo = await serviceBusManagementClient.getNamespaceProperties();
+  const namespaceInfo = await ServiceBusAdministrationClient.getNamespaceProperties();
   console.log("Type of the namespace - ", namespaceInfo.namespaceType);
 
-  await serviceBusManagementClient.deleteQueue(queueName);
-  const queueExists = await serviceBusManagementClient.queueExists(queueName);
+  await ServiceBusAdministrationClient.deleteQueue(queueName);
+  const queueExists = await ServiceBusAdministrationClient.queueExists(queueName);
   if (queueExists == true) {
     console.log("Something went wrong, queue should have been deleted");
     return;